api/requirements.txt
# Core Framework
fastapi==0.104.1
python-multipart==0.0.6
uvicorn==0.24.0

# PDF Processing - using latest version with prebuilt wheels
PyMuPDF>=1.24.0

# OCR - MVP
pytesseract==0.3.10
Pillow==10.1.0

# OCR - Production (via API calls)
requests==2.31.0

# Utilities
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2
faker==21.0.0
